Initial release
With the go to release the code source in Apache 2.0 Licence, we removed
the history to hide some secrets during the internal development.

Everything which was specific to YP will be redesigned from this point to have an independant code

There is a lot of reference to drone but this counter is now independant of the drone solution since we removed
the legacy support which used a direct MySQL connection to the drone database.