Customizable Config
Config class can now be overriden and passed to the application.
This is a more generic approach to avoid code fork for specific cases.

app migrate to a class App
settings include an init function called from App

config and storage are still set global in the settings module but config can be passed
to the settings module.

Config is simplified and more generic.
The Storage backend will not have a direct access to the config anymore as this is not rquired.